backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.15.1
pydantic==2.9.2
pydantic-settings==2.5.2
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1